\name{actors}
\alias{actors}
\title{List actor codes}
\usage{
  actors(edo)
}
\arguments{
  \item{edo}{Event data}
}
\value{
  Array of actor codes
}
\description{
  Lists actor codes
}
\details{
  Lists all the actor codes that occur in the event data in
  alphabetical order.
}
\author{
  Will Lowe
}
\seealso{
  \code{\link{sources}}, \code{\link{targets}},
  \code{\link{codes}}
}

