\name{add_eventscale}
\alias{add_eventscale}
\title{Apply eventscale to event data}
\usage{
  add_eventscale(edo, sc)
}
\arguments{
  \item{edo}{Event data}

  \item{sc}{scale}
}
\value{
  Event data with a scaling
}
\description{
  Applies an eventscale to event data
}
\details{
  Applies an eventscale to event data.  This adds a new
  field in the event data with the same name as the
  eventscale.  Add as many as you want to keep around.
}
\author{
  Will Lowe
}

