% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{as.data.frame.eventscale}
\alias{as.data.frame.eventscale}
\title{Coerce Event Scale to Data Frame}
\usage{
\method{as.data.frame}{eventscale}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an event scale}

\item{row.names}{ignored}

\item{optional}{ignored}

\item{...}{ignored

This function converts a list with event codes as names and event scores 
as values into a data frame with column 'code' containing the event codes
and column 'score' as the event's score}
}
\value{
a data.frame containing event codes and scores
}
\description{
Coerce Event Scale to Data Frame
}
