\docType{data}
\name{balkans.weis}
\alias{balkans.weis}
\title{Balkans conflict events in WEIS encoding}
\description{
  Event data on the conflict during the collapse of
  Yugoslavia.  Events are coded according to an extended
  WEIS scheme by the KEDS Project.
}
\author{
  KEDS Project
}
\references{
  \url{http://web.ku.edu/~keds/data.dir/balk.html}
}
\keyword{data}
\keyword{datasets}

