% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgge.R
\name{bgge}
\alias{bgge}
\alias{rbgge}
\alias{pbgge_marginal_x}
\title{\strong{B}ivariate Distribution with \strong{G}amma and \strong{G}eneralized
\strong{E}xponential Marginals}
\usage{
rbgge(n, m, beta, rounding = FALSE)

pbgge_marginal_x(q, n, beta, lower.tail = TRUE)
}
\arguments{
\item{n}{when provided to rsmp, this is the number of observations. Otherwise
it represents the duration of an event.}

\item{m}{the number of elements in the vectors. This is expected to either be
a single number (all generated vectors have the same number of elements) or a
vector of length n (in which case the number of elements in each vector is
specified)}

\item{beta}{numeric > 0; the rate parameter of the exponential distribution
which describes individual exceedances. It is the reciprocal of the mean.}

\item{rounding}{logical; if TRUE, generated Pareto II values are rounded to
the nearest whole number. This feature is present in case this is important
for a comparison of actual observations to some theoretical output.}

\item{q}{a number; value given univariate functions of univariate marginals}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}
}
\value{
rbgge generates random values of the distribution and returns a them
data.frame() with one obervation per row and three columns:
\itemize{
\item n the number of elements in the IID Expoential vector
\item x the sum of the elements in the vector
\item y the maximal value in the vector
}
}
\description{
rbgge generates sums and maxima of IID exponential vectors of specified
lengths
}
\details{
pbgge evaulates the marginal distribution of the sum of the exponential
components, called "magnitudes"; This marginal distribution is merely an
Erlang (Gamma) distribution with shape n and rate beta
}
\examples{
rbgge(n = 3, m = c(3,4,2), beta = 1/100)
}
