% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\docType{data}
\name{cameo.scale}
\alias{cameo.scale}
\title{CAMEO codes to conflict-cooperation scale}
\description{
A mapping of CAMEO event codes to [-10,10] representing a scale
of conflict and cooperation, developed by the KEDS project.
Taken from the documentation of the KEDS_Count software.
}
\details{
The version of CAMEO used here is 0.9B5 [07.03.2021].
}
\references{
Parus Analytics: https://eventdata.parusanalytics.com/data.dir/cameo.html
}
\author{
KEDS Project
}
\keyword{data}
