% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\docType{data}
\name{casia.weis}
\alias{casia.weis}
\title{Central Asia events with WEIS event coding}
\description{
Event data on Central Asia.  Events are
coded according to the WEIS scheme by the KEDS Project.  The event
stream contains 8377 events occurring between 02/05/1989 and 31/07/1999 
involving 152 sources and 157 targets.
}
\details{
Original data comes from file CASIA.LEADS.6CODE (six character actor codes and coded from leads) 
with duplicates removed using the one_a_day filter.
}
\references{
\url{https://www.parusanalytics.com/eventdata/data.dir/casia.html}
}
\author{
KEDS Project
}
\keyword{data}
