% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{codes}
\alias{codes}
\title{List event codes}
\usage{
codes(edo)
}
\arguments{
\item{edo}{Event data}
}
\value{
Array of event codes
}
\description{
Lists event codes
}
\details{
Lists all the event codes that appear in the event data
}
\seealso{
\code{\link{sources}}, \code{\link{targets}}, \code{\link{actors}}
}
\author{
Will Lowe
}
<<<<<<< HEAD
=======
\seealso{
\code{\link{sources}}, \code{\link{targets}}, \code{\link{actors}}
}

>>>>>>> 04e7b17b1e0e1dc576ccecff1ee015c5e8a3d0e2
