% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_events.R
\name{create_events}
\alias{create_events}
\alias{create_events_from_vector}
\alias{create_events_from_df}
\title{Create events from observed time series and a threshold}
\usage{
create_events_from_vector(v, threshold)

create_events_from_df(
  df,
  event_var = "precip",
  thresh_col = "threshold",
  time_col = "date",
  aux_vars = c("tmax", "tmin"),
  aux_funs = list(max = max, min = min, sum = sum, mean = mean)
)
}
\arguments{
\item{v}{A vector (assumed to be numeric) representing some variable of
interest which has been measured at regular time intervals; Missing values
or "gaps in observed series" must be explicitly filled with NAs.}

\item{threshold}{A numeric threshold which defines the minimum value for
which an observation would be considered part of an event}

\item{df}{A data.frame of observations recorded at regular intervals (
restricted, for now, to days).}

\item{event_var}{(character) the variable within df for which events are
defined, e.g. "precip"}

\item{thresh_col}{(character)the name of the column containing a
pre-calcuated threshold}

\item{time_col}{(character) the name of the column with time information;
"date" by default, but it could also be something like "date_time_pacific"
if finer-res data are to be considered}

\item{aux_vars}{(character vector) other variables within df which you want
to summarize alongside events. For example, "tmin" or some measure of
stability might be of interest during precipitation events}

\item{aux_funs}{(list of functions) summary functions to to apply to,
aux_vars such as min or max; Internally, lists of functions are applied to
multiple aux_vars by means of the dplyr package's mutate/summarize-across
construct. See documentation for dplyr::across for details.}
}
\value{
A list containing:
\itemize{
\item threshold, the vector above which the series was truncated in order to
define events
\item event_summaries_df, a data.frame() -
Each row provides a basic summary of events, including their starting
indices, durations, magnitudes, maxima, and whether or not the duration,
magnitude, and maximum are to be taken at face value}
}
\description{
Create events from observed time series and a threshold
}
\examples{
v <- c(333,587,0,201,82,225,70,62,47,101,187,NA,NA,126,222,86,NA,125)
# ^^ perhaps precipitation measured in 10ths of mm, with some missing values
# in the mix
create_events_from_vector(v, 100)

df <- data.frame(
precip = v,
tmax = rnorm(n = 18, mean = 10, sd = 2),
threshold = 100,
date = seq.Date(from = as.Date("2025-01-01"), by = "day", length.out = 19)[-10])

event_var = "precip"
thresh_col = "threshold"
time_col = "date"

aux_vars = c("tmax", "tmin")
aux_funs = list(max = max, min = min, sum = sum, mean = mean)

create_events_from_df(df = df, event_var = "precip",
aux_vars = aux_vars, aux_funs = aux_funs,
thresh_col = thresh_col, time_col = time_col)


\dontrun{
$threshold
[1] 100

$event_summaries_df
  start_index length magnitude maximum length_uncertain
1           1      2       487     487             TRUE
2           4      1       101     101            FALSE
3           6      1       125     125            FALSE
4          10      2        88      87             TRUE
5          14      2       148     122             TRUE
6          18      1        25      25             TRUE
}
}
