% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_events_utils.R
\name{create_events_utils}
\alias{create_events_utils}
\alias{df_complete_time_column}
\alias{threshold_given_desc}
\alias{parse_threshold_desc}
\title{Data prep utils used by create_events\* functions}
\usage{
df_complete_time_column(df, time_colname = "date", by = "day")

threshold_given_desc(v, threshold_desc)

parse_threshold_desc(threshold_desc)
}
\arguments{
\item{df}{data.frame() with a column called time_colname with Date or POSIXt
entries}

\item{time_colname}{(character) the name of the date or time column}

\item{by}{(character) the increment of the sequence, e.g. "day" or "hour";
See base::seq.Date or base::seq.POSIXt for details}
}
\description{
Data prep utils used by create_events\* functions
}
\examples{
df <- data.frame(date = as.Date(c("2025-01-01", "2025-01-03")), precip = c(0,1.2))
df_complete_time_column(df)

}
