% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_utils.R
\name{determine_stopping_point}
\alias{determine_stopping_point}
\title{Find a quantile of a discrete distribution to achieve a low exceedance probability}
\usage{
determine_stopping_point(
  event_duration_marginal_pmf,
  event_duration_marginal_pmf_args,
  tol = 10^(-10),
  max_N = 10 * (2^8)
)
}
\arguments{
\item{event_duration_marginal_pmf}{a function; the marginal CDF for event durations}

\item{event_duration_marginal_pmf_args}{a list; arguments for the marginal duration CDF}

\item{tol}{a very small number; minimal exceedance probability desired}

\item{max_N}{a number; highest N to search}
}
\value{
an integer
}
\description{
This is a utility function used when calculating marginal distributions of
magnitudes and maxima of trivariate events. These marginal distributions are
generally calculated as sums of CDFs conditioned on duration multiplied by
probabilities of durations. Distributions of durations of all events we
anticipate seeing will be decreasing after a certain duration. This function
finds the duration we need to calculate the marginal probabilities to an
arbitrary precision (the "tol").
}
\examples{
determine_stopping_point(pgeom, list(prob = 0.1), max_N = 200)
}
