% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgeom.R
\name{dhgeom}
\alias{dhgeom}
\title{The Hurdle-geometric Distribution}
\usage{
dhgeom(x, prob_p, prob_q, log = FALSE)
}
\arguments{
\item{log}{}
}
\description{
Density, distribution function, quantile function and random generation for
the 1-inflated geometric distribution with parameters prob_p and prob_q.
}
