% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdiscretelomax.R
\name{discretelomax}
\alias{discretelomax}
\alias{ddiscretelomax}
\alias{rdiscretelomax}
\alias{pdiscretelomax}
\alias{qdiscretelomax}
\title{Discrete Lomax (Pareto II) Distribution}
\usage{
ddiscretelomax(x, dlomax_alpha, dlomax_prob_p, return_type = "vector")

rdiscretelomax(n, dlomax_alpha, dlomax_prob_p)

pdiscretelomax(
  q,
  dlomax_alpha,
  dlomax_prob_p,
  lower.tail = TRUE,
  log.p = FALSE,
  return_type = "vector"
)

qdiscretelomax(
  p,
  dlomax_alpha,
  dlomax_prob_p,
  lower.tail = TRUE,
  return_type = "vector"
)
}
\arguments{
\item{dlomax_prob_p}{}
}
\description{
Discrete Lomax (Pareto II) Distribution
}
\details{
In the output vector, it is the prop_p which vary more quickly. If you
provide 2 dlomax_alphas and 3 dlomax_prob_ps, the first 3 numbers in the output are
associated with the first dlomax_alpha.
}
\examples{
\dontrun{
ddiscretelomax(x = 1:10, dlomax_alpha = c(0,0.2,0.4),
prob = c(0.6), return_type = "data.frame") \%>\%
ggplot2::ggplot(
ggplot2::aes(x = x, y = prob_mass, group = dlomax_alpha, fill = dlomax_alpha)) +
ggplot2::geom_col(position = "dodge") +
ggplot2::scale_fill_binned() +
ggplot2::scale_y_continuous(trans = "log")

pdiscretelomax(q = seq(1:15),
dlomax_alpha = c(0,0.05,0.1,0.2),
dlomax_prob_p = c(0.5),
return_type = "data.frame", lower.tail = F)  \%>\%
ggplot2::ggplot(ggplot2::aes(x = q,
y = non_exceedance_prob, group = dlomax_alpha, color = dlomax_alpha)) +
ggplot2::geom_path() +
ggplot2::scale_x_continuous(breaks = 1:15) +
ggplot2::scale_y_continuous(trans = "log10")

qdiscretelomax(p = seq(0.54,0.99, by = 0.05),
dlomax_alpha = c(0,0.1,0.3),
dlomax_prob_p = 0.5,
return_type = "data.frame") \%>\%
ggplot2::ggplot(ggplot2::aes(x = lte_prob,
y= q, color = dlomax_alpha)) +
ggplot2::geom_point() +
ggplot2::lims(y = c(0,NA))
}
}
