% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events-package.R
\docType{package}
\name{events-package}
\alias{events}
\alias{events-package}
\title{events: Tools for identifying, analyzing, and plotting events from time series}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alexander Weyant \email{aweyant@ucsd.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
