% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\docType{package}
\name{events}
\alias{events}
\alias{package-events}
\title{Stores and manipulates event data}
\description{
Stores, manipulates, scales, aggregates and creates directed dyadic time series
from event data generated by KEDS, TABARI, or any other extraction tool 
with similarly structured output.
}
\details{
Events offers simple methods for aggregating and renaming actors and event codes, 
applying event scales, and constructing regular time series at a choice of
temporal scales and measurement levels.
}
\author{
Will Lowe \email{conjugateprior@gmail.com}
}
