% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{filter_codes}
\alias{filter_codes}
\title{Discard all but relevant event codes}
\usage{
filter_codes(edo, fun = function(x) {     return(TRUE) })
}
\arguments{
\item{edo}{Event data}

\item{fun}{Function that returns \code{TRUE} for event codes that should not be discarded, or single event code, or a vector of them.}
}
\value{
Event data containing only events that pass through \code{fun}
}
\description{
Discards all but relevant event codes
}
\details{
Applies the filter function to each event code to see whether
to keep the observation.
}
\author{
Will Lowe
}
\seealso{
\code{\link{filter_actors}}, \code{\link{filter_time}}
}

