% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{filter_eventdata}
\alias{filter_eventdata}
\title{Filter events data}
\usage{
filter_eventdata(edo, fun, which)
}
\arguments{
\item{edo}{Events data object}

\item{fun}{Function that should be applied}

\item{which}{Which fields should be filtered}
}
\value{
Event data
}
\description{
Applies a generic field filter to event data
}
\details{
This function applies a filter function to event data.
It is the workhorse function behind the \code{filter_} functions.
You should use these in ordinary use.
}
\examples{
data(levant.cameo)
sp <- spotter("PAL", "ISR")
ev_targ <- filter_eventdata(levant.cameo, sp, "target") # these actors as targets
head(ev_targ, 3)
ev_dyad <- filter_eventdata(levant.cameo, sp, c("source", "target")) # source and target
head(ev_dyad, 3)
}
\author{
Will Lowe
}
