% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{filter_time}
\alias{filter_time}
\title{Restrict events to a time period}
\usage{
filter_time(edo, start = min(edo$date), end = max(edo$date))
}
\arguments{
\item{edo}{Event data}

\item{start}{Something convertable to a \code{Date} object}

\item{end}{Something convertable to a \code{Date} object}
}
\value{
Event data restricted to a time period
}
\description{
Restricts events to a time period
}
\details{
Restricts events on or after \code{start} and on or before \code{end}.
<<<<<<< HEAD
=======
}
\author{
Will Lowe
>>>>>>> 04e7b17b1e0e1dc576ccecff1ee015c5e8a3d0e2
}
\seealso{
\code{\link{filter_codes}}, \code{\link{filter_actors}}
}
\author{
Will Lowe
}
