% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpsmp.R
\name{hdpsmp}
\alias{hdpsmp}
\alias{phdpsmp}
\alias{phdpsmpdt}
\alias{qhdpsmpdt}
\alias{cdf_hdpsmp}
\title{\strong{H}urdle \strong{D}iscrete \strong{P}areto \strong{S}ums & \strong{M}axima of \strong{P}areto II Vectors}
\usage{
phdpsmp(
  x_lower = 0,
  x_upper = Inf,
  y_lower = 0,
  y_upper = Inf,
  n_lower = 1,
  n_upper = Inf,
  alpha,
  beta,
  prob_q,
  dlomax_prob_p,
  dlomax_alpha
)

phdpsmpdt(
  q,
  threshold,
  alpha,
  beta,
  prob_q,
  dlomax_prob_p,
  dlomax_alpha,
  lower.tail = TRUE,
  log = FALSE
)

qhdpsmpdt(
  p,
  threshold,
  alpha,
  beta,
  prob_q,
  dlomax_prob_p,
  dlomax_alpha,
  lower.tail = TRUE,
  log = FALSE
)

cdf_hdpsmp(x, y, n, alpha, beta, prob_q, dlomax_prob_p, dlomax_alpha)
}
\arguments{
\item{x_lower, x_upper, y_lower, y_upper}{each are individual numbers; limits of
integration of the SMP CDF}

\item{n_lower, n_upper}{each are individual numbers; limits of summation of
for the duration of the a trivariate distribution, where n_lower is the
lowest duration included in the summation and n_upper is the largest included
in the summation; for example, n_lower = 3, n_upper = 5 sums over {3,4,5},
whereas n_lower = 3, n_upper = 3 sums over {3}.}

\item{alpha, beta}{tail/shape parameter and scale parameters of a Multivariate
Pareto II distribution (see details for the parameterization used).}

\item{prob_q}{probability of success in the first trial. \eqn{0 \le} prob_q \eqn{\le 1}}

\item{q}{a number; value given univariate functions of univariate marginals}

\item{threshold}{A numeric threshold which defines the minimum value for
which an observation would be considered part of an event}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{x, y}{numbers; magnitude and maxima of a Pareto II vector}

\item{n}{when provided to rsmp, this is the number of observations. Otherwise
it represents the duration of an event.}
}
\description{
Probability and random generation functions for the trivariate HDPSMP distribution
}
