% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgeom.R
\name{hgeom}
\alias{hgeom}
\alias{dhgeom}
\alias{phgeom}
\alias{qhgeom}
\alias{rhgeom}
\title{The Hurdle-geometric Distribution}
\usage{
hgeom(x, q, p, n, prob_p, prob_q, log, log.p, lower.tail)

dhgeom(x, prob_p, prob_q, log = FALSE)

phgeom(q, prob_p, prob_q, lower.tail = TRUE, log.p = FALSE)

qhgeom(p, prob_p, prob_q, lower.tail = TRUE, log.p = FALSE)

rhgeom(n, prob_p, prob_q)
}
\arguments{
\item{x, q}{vector of quantiles representing the number of failures in a
sequence of Bernoulli trials before success occurs.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be
the number required.}

\item{prob_p}{probability of success in each trial after the first trial, if
the first trial was not a success. \eqn{0 <} prob_p \eqn{\le 1}.}

\item{prob_q}{probability of success in the first trial. \eqn{0 \le} prob_q \eqn{\le 1}}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}
}
\value{
hgeom is a dummy function to keep the hurdle geometric documentation
together.

dhgeom gives the density, phgeom gives the distribution function, qhgeom
gives the quantile function, and rhgeom generates random deviates.

Invalid prob_p or prob_q will result in return value NaN, with a warning.

rhgeom returns a vector of type integer unless generated values exceed the
maximum representable integer when double values are returned since R
version 4.0.0.
}
\description{
Density, distribution function, quantile function and random generation for
the 1-inflated geometric distribution with parameters prob_p and prob_q.
}
\examples{
plot(dhgeom(x = 1:25, prob_p = 0.3, prob_q = 0.6))
}
