% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janky_mle_discretelomax.R
\name{janky_mle_discretelomax}
\alias{janky_mle_discretelomax}
\title{Direct maximum likelihood estimation of the parameters of a discrete Lomax
distribtuion}
\usage{
janky_mle_discretelomax(v)
}
\arguments{
\item{v}{a vector of integers thought to be drawn from a discrete Lomax
distribution}
}
\value{
A list with elements:
\itemize{
\item estimate - a vector of estimates corresponding to each parameter in the
distribution
}
}
\description{
Direct maximum likelihood estimation of the parameters of a discrete Lomax
distribtuion
}
\examples{
sample <- rdiscretelomax(n = 100, dlomax_alpha = 0.01, dlomax_prob_p = 0.5)
janky_mle_discretelomax(sample)
}
