% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janky_mle_hgeom.R
\name{janky_mle_hgeom}
\alias{janky_mle_hgeom}
\title{Direct maximum likelihood estimation of the parameters of a hurdle-geometric
distribution}
\usage{
janky_mle_hgeom(v)
}
\arguments{
\item{v}{a vector of integers thought to be drawn from a hurdle-geometric
distribution}
}
\value{
A list with elements:
\itemize{
\item estimate - a vector of estimates corresponding to each parameter in the
distribution
}
}
\description{
Direct maximum likelihood estimation of the parameters of a hurdle-geometric
distribution
}
\examples{
sample_n <- rhgeom(n = 20, prob_p = 0.3, prob_q = 0.7)
janky_mle_hgeom(.Last.value)
}
