% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janky_mle_hgsmp.R
\name{janky_mle_hgsmp}
\alias{janky_mle_hgsmp}
\alias{alpha_est}
\alias{beta_est}
\title{Maximum Likelihood Estimator Calculation for the Hurdle-Geometric Sums &
Maxima of Pareto Vectors distribution}
\usage{
janky_mle_hgsmp(magnitudes, durations)

alpha_est(magnitudes, durations, num_search = TRUE)

beta_est(magnitudes, durations, alpha = NULL)
}
\arguments{
\item{magnitudes}{sums of Pareto II vectors}

\item{durations}{number of elements in Pareto II vectors}

\item{num_search}{logical; whether or not to initiate a numerical search for
alpha. This is mostly for internal use.}

\item{alpha}{optimal alpha used to calculate beta. This is only used by beta_est()
because beta's optimal value depends on the optimal value of for alpha. If you
estimate both alpha and beta, there is no reason to repeat the first step.}
}
\value{
A list with elements:
\itemize{
\item estimate raw ML estimates of the four HGSMP parameters}
}
\description{
Maximum Likelihood Estimator Calculation for the Hurdle-Geometric Sums &
Maxima of Pareto Vectors distribution
}
\section{Functions}{
\itemize{
\item \code{alpha_est()}: Helper function to estimate alpha

\item \code{beta_est()}: Helper function to estimate beta

}}
\examples{
random_events_df <- rsmp(n = 100, m = c(rep(1,70), rep(2,22), rep(3,7), 4),
alpha = 0.03, beta = 1/100)
janky_mle_hgsmp(random_events_df$x, random_events_df$n)
}
