% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\docType{data}
\name{levant.cameo}
\alias{levant.cameo}
\title{Levant events with CAMEO event coding}
\description{
Event data on Middle East.  Events are
coded according to the CAMEO scheme by the KEDS Project.  The event
stream contains 145709 events occurring between 15/04/1979 and 30/11/2011 
involving 741 sources and 688 targets.
}
\details{
Original data comes from file REULE.201111.evt with documentation lines
(marked as DOC DOC 999), match information, and duplicates removed.
}
\references{
\url{http://www.parusanalytics.com/eventdata/data.dir/levant.html}
}
\author{
KEDS Project
}
\keyword{data}
