% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{map_codes}
\alias{map_codes}
\title{Aggregate event codes}
\usage{
map_codes(edo, fun = function(x) {     return(x) })
}
\arguments{
\item{edo}{Event data}

\item{fun}{Function specifying the aggregation mapping}
}
\value{
Event data with new event codes
}
\description{
Aggregates event codes
}
\details{
This function relabels event codes according to \code{fun},
The \code{mapper} function is an easy way to make such a function.

This can also be used as a renaming function, but it is most
useful when multiple codes should be treated as equivalent.
}
\seealso{
\code{\link{map_actors}}
}
\author{
Will Lowe
}
