% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{plot.scaled_directed_dyad}
\alias{plot.scaled_directed_dyad}
\title{Plot scaled directed dyad}
\usage{
\method{plot}{scaled_directed_dyad}(x, ...)
}
\arguments{
\item{...}{Extra arguments to plot}

\item{dyad}{One directed dyadic time series from the \code{make_dyads} function}
}
\value{
Nothing, used for side effect
}
\description{
Plots scaled directed dyad
}
\details{
A convenience function to plot the named scale within a directed dyad against time.
}
\author{
Will Lowe
}

