\name{plot_dyad}
\alias{plot_dyad}
\title{Plot scaled directed dyad}
\usage{
  plot_dyad(dyad, ...)
}
\arguments{
  \item{dyad}{One directed dyadic time series from the
  \code{make_dyads} function}

  \item{...}{Extra arguments to plot}
}
\value{
  Nothing, used for side effect
}
\description{
  Plots scaled directed dyad
}
\details{
  A convenience function to plot the named scale within a
  directed dyad against time.
}
\author{
  Will Lowe
}

