\name{print.eventdata}
\alias{print.eventdata}
\title{Show the first few events in an event data set}
\usage{
\method{print}{eventdata}(edo, ...)
}
\arguments{
  \item{edo}{Event data}
}
\value{
Silently returns the event data set
}
\description{
Prints out the first few events of an event data set
}
\author{
Will Lowe
}

