% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_utils.R
\name{psdeddt}
\alias{psdeddt}
\title{its name cannot be anticipated
corresponding event magnitude}
\usage{
psdeddt(
  q,
  length = 1,
  threshold,
  event_magnitude_conditional_cdf,
  event_magnitude_conditional_cdf_args,
  event_length_arg_name = "n",
  lower.tail = TRUE,
  return_NA_outside_support = FALSE
)
}
\description{
its name cannot be anticipated
corresponding event magnitude
}
