% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_petrarch}
\alias{read_petrarch}
\title{Reads Petrarch event data output files}
\usage{
read_petrarch(d, use.quad = FALSE, use.score = FALSE)
}
\arguments{
\item{d}{name of a tsv containing events, or a list of such names}
}
\value{
An event data set
}
\description{
Reads Petrach output files and retains only the event date, the 
main source and target actor codes, and the CAMEO and quad score 
of the event.
}
\details{
Tested on the data sets at \url{http://phoenixdata.org/data}.
Output column names are 'date', 'source', 'target', 'code' and 'score', if use.score is TRUE
}
\author{
Will Lowe
}

