% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlomax.R
\name{rmvlomax}
\alias{rmvlomax}
\title{Multivariate Pareto II Vectors}
\usage{
rmvlomax(n, m, alpha, beta, rounding = FALSE)
}
\arguments{
\item{n}{when provided to rsmp, this is the number of observations. Otherwise
it represents the duration of an event.}

\item{m}{the number of elements in the vectors. This is expected to either be
a single number (all generated vectors have the same number of elements) or a
vector of length n (in which case the number of elements in each vector is
specified)}

\item{alpha, beta}{tail/shape parameter and scale parameters of a Multivariate
Pareto II distribution (see details for the parameterization used).}

\item{rounding}{logical; if TRUE, generated Pareto II values are rounded to
the nearest whole number. This feature is present in case this is important
for a comparison of actual observations to some theoretical output.}
}
\value{
A list of random Pareto II vectors
}
\description{
Multivariate Pareto II Vectors
}
\examples{
rmvlomax(n = 3, m = c(3,4,2), alpha = 0.03, beta = 1/100)
\dontrun{
[[1]]
[1] 170.56225 120.49477  14.40258

[[2]]
[1] 192.361464 274.643827  64.666721   9.729169

[[3]]
[1] 108.0837 255.7757
}
}
