\name{scale_codes}
\alias{scale_codes}
\title{Show which events are scaleable}
\usage{
  scale_codes(es)
}
\arguments{
  \item{es}{Eventscale}
}
\value{
  Array of scaleable event codes
}
\description{
  Shows which events codes are covered by a scale
}
\details{
  Returns an array of event codes to which an eventscale
  assigns a value.
}
\author{
  Will Lowe
}

