% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smp.R
\name{smp}
\alias{smp}
\alias{rsmp}
\alias{psmp}
\alias{psmp_marginal_x}
\alias{psmp_marginal_y}
\alias{cdf_smp}
\title{The Distribution of Sums & Maxima of Multivariate Pareto II Vectors}
\usage{
rsmp(n, alpha, beta, m, rounding = FALSE)

psmp(
  x_lower = (1e-10)/beta,
  x_upper = Inf,
  y_lower = (1e-10)/beta,
  y_upper = Inf,
  n = 1,
  alpha,
  beta,
  x_lims = NULL,
  y_lims = NULL
)

psmp_marginal_x(q, n, alpha, beta, lower.tail = TRUE)

psmp_marginal_y(q, n, alpha, beta, lower.tail = TRUE)

cdf_smp(x, y, n, alpha, beta)
}
\arguments{
\item{n}{when provided to rsmp, this is the number of observations. Otherwise
it represents the duration of an event.}

\item{alpha, beta}{tail/shape parameter and scale parameters of a Multivariate
Pareto II distribution (see details for the parameterization used).}

\item{m}{the number of elements in the vectors. This is expected to either be
a single number (all generated vectors have the same number of elements) or a
vector of length n (in which case the number of elements in each vector is
specified)}

\item{rounding}{logical; if TRUE, generated Pareto II values are rounded to
the nearest whole number. This feature is present in case this is important
for a comparison of actual observations to some theoretical output.}

\item{x_lower, x_upper, y_lower, y_upper}{each are individual numbers; limits of
integration of the SMP CDF}

\item{x_lims, y_lims}{each are vectors of length 2; an alternative way to
specify the limits of integration as vectors of upper and lower bounds}

\item{q}{a number; value given univariate functions of univariate marginals}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{x, y}{numbers; magnitude and maxima of a Pareto II vector}
}
\value{
rsmp generates random values of the distribution and returns a them
data.frame() with one obervation per row and three columns:
\itemize{
\item n the number of elements in the Pareto II vector
\item x the sum of the elements in the Pareto II vector
\item y the maximal value in the Pareto II vector
}

smp integrates the CDF between user-provided limits. The result is returned
as a number in [0,1]
}
\description{
Density, distribution function, quantile function and random generation for
the Sums \& Maxima of Pareto II Vectors of specified lengths
}
\examples{
rsmp(n = 3, m = c(3,4,2), alpha = 0.03, beta = 1/100)
\dontrun{
  n        x        y
1 3 241.6733 157.2649
2 4 357.7778 187.8859
3 2 219.1081 140.8779
}
psmp(x_upper = 100, y_lower = 40, n = 3, alpha = 0.03, beta = 1/100)

psmp(x_upper = 100, n = 3, alpha = 0.03, beta = 1/100)
# This should provide the same output as...
psmp_marginal_x(q = 100, n = 3, alpha = 0.03, beta = 1/100)

psmp_marginal_y(q = 100, n = 3, alpha = 0.03, beta = 2/100)
cdf_smp(x = 200, y = 100, n = 3, alpha = 0.03, beta = 2/100)
}
