% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{sources}
\alias{sources}
\title{List source actor codes}
\usage{
sources(edo)
}
\arguments{
\item{edo}{Event data}
}
\value{
Array of actor codes
}
\description{
Lists source actor codes
}
\details{
Lists all the actor codes that appear as a source in the event data
in alphabetical order.
}
\seealso{
\code{\link{actors}}, \code{\link{targets}}, \code{\link{codes}}
}
\author{
Will Lowe
}
