% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_utils.R
\name{ted}
\alias{ted}
\alias{pteddt}
\alias{psdeddt}
\title{\strong{P}robability of \strong{T}rivariate \strong{E}vent \strong{D}istribution \strong{D}erived
\strong{T}otal (P TED DT) and \strong{P}robability of \strong{S}et \strong{D}uration \strong{E}vent
\strong{D}istribution \strong{D}erived \strong{T}otal}
\usage{
pteddt(
  q = NULL,
  threshold,
  event_duration_marginal_pmf,
  event_duration_marginal_pmf_args = NULL,
  event_length_arg_name = "n",
  event_magnitude_conditional_cdf,
  event_magnitude_conditional_cdf_args,
  lower.tail = TRUE,
  tol = 10^(-6),
  max_N = 100
)

psdeddt(
  q = NULL,
  length = 1,
  threshold,
  event_magnitude_conditional_cdf,
  event_magnitude_conditional_cdf_args,
  event_length_arg_name = "n",
  lower.tail = TRUE,
  return_NA_outside_support = FALSE
)
}
\arguments{
\item{q}{a number; the event total, including the below-threshold portion}

\item{threshold}{a number; the minimum threshold used to define events}

\item{event_duration_marginal_pmf}{a function; the marginal CDF for event durations}

\item{event_duration_marginal_pmf_args}{a list; arguments for the marginal duration CDF}

\item{event_length_arg_name}{a string; the name of the argument used for the
duration of an event in the CDF of the magnitude conditioned duration. For
example, this would be "n" for psmp_marginal_x()}

\item{event_magnitude_conditional_cdf}{a function; the CDF of event magnitudes
conditioned on durations, e.g. psmp_marginal_x()}

\item{event_magnitude_conditional_cdf_args}{a list; other arguments, such as
parameters of the distribution}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{tol}{a very small number; minimal exceedance probability desired}

\item{max_N}{a number; highest N to search}

\item{length}{integer; the duration of the event}

\item{return_NA_outside_support}{logical; should an unsupported event cause
a return of NA or 0? Decide here.}
}
\value{
a number in the interval [0,1], the probability of not exceeding or
exceeding q
}
\description{
This evaluates the CDF of the total accumulation of an amount over an event.
}
\details{
pteddt() by way of example:
a 3 day precipitation event might have a magnitude of 430m. If
the threshold for defining the event was 75mm/day, then the actual total
precipitation over this event was 655mm [430 + 3*75].

Note that durations and daily accumulations are, in general, both random
variables, so if we are interested in probabilties (return intervals) of
uninterrupted precipitation accumulations, we must construct an event total
CDF from (at least) two components: the (discrete) PMF describing event
durations and the CDF for magnitudes, conditioned upon a duration. The Law of
Total Probability lets us create a CDF for the event total, which is quite
meaningful if we are studying precipitation.

psdeddt() evaluates the CDF for event totals of a given duration (hence, "set
duration").
}
\examples{
q = 100; length = 2; threshold = 20
alpha = 0.02; beta = 0.04
prob_p = 0.6; prob_q = 0.7
event_length_arg_name = "n"

pteddt(q = q,
threshold = threshold,
event_duration_marginal_pmf = dhgeom,
event_duration_marginal_pmf_args = list(prob_p = prob_p, prob_q = prob_q),
event_length_arg_name = event_length_arg_name,
event_magnitude_conditional_cdf = psmp_marginal_x,
event_magnitude_conditional_cdf_args = list(alpha = alpha, beta = beta),
lower.tail = TRUE,
tol = 10^(-6),
max_N = 100)
rsmp(n = 3, m = c(3,4,2), alpha = 0.03, beta = 1/100)
\dontrun{
  n        x        y
1 3 241.6733 157.2649
2 4 357.7778 187.8859
3 2 219.1081 140.8779
}
}
