\name{actor_interactions}
\alias{actor_interactions}
\title{cross-tabulate source and target actor interaction counts}
\usage{
actor_interactions(edo)
}
\arguments{
  \item{edo}{Event data}
}
\value{
Table of actor interaction counts
}
\description{
Cross-tabulate actor interactions
}
\details{
A cross-tabulation of events in terms of their source and
target actors.
}
\author{
Will Lowe
}

