\docType{data}
\name{balkans.weis}
\alias{balkans.weis}
\title{Balkans conflict events in WEIS encoding}
\description{
  Event data on the conflict during the collapse of
  Yugoslavia.  Events are coded according to an extended
  WEIS scheme by the KEDS Project.  The event stream
  contains 72953 events occurring between 2 April 1989 and
  31 July 2003 involving 325 actors.
}
\author{
  KEDS Project
}
\references{
  \url{http://web.ku.edu/~keds/data.dir/balk.html}
}
\keyword{data}
\keyword{dataset}

