\docType{data}
\name{cameo.scale}
\alias{cameo.scale}
\title{CAMEO codes to conflict-cooperation scale}
\description{
  A mapping of CAMEO event codes to [-10,10] representing a
  scale of conflict and cooperation, developed by the KEDS
  project. Taken from the documentation of the KEDS_Count
  software.
}
\details{
  The version of CAMEO used here is 0.9B5 [08.04.15].
}
\author{
  KEDS Project
}
\references{
  \url{http://web.ku.edu/~keds/}
}
\keyword{data}
\keyword{dataset}

