% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{codes}
\alias{codes}
\title{List event codes}
\usage{
codes(edo)
}
\arguments{
\item{edo}{Event data}
}
\value{
Array of event codes
}
\description{
Lists event codes
}
\details{
Lists all the event codes that appear in the event data
}
\examples{
data(levant.cameo)
cod <- codes(levant.cameo)
head(codes)
tail(codes)
}
\seealso{
\code{\link{sources}}, \code{\link{targets}}, \code{\link{actors}}
}
\author{
Will Lowe
}
