\name{codes}
\alias{codes}
\title{List event codes}
\usage{
  codes(edo)
}
\arguments{
  \item{edo}{Event data}
}
\value{
  Array of event codes
}
\description{
  Lists event codes
}
\details{
  Lists all the event codes that appear in the event data
}
\author{
  Will Lowe
}
\seealso{
  \code{\link{sources}}, \code{\link{targets}},
  \code{\link{actors}}
}

