% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_events.R
\name{create_events}
\alias{create_events}
\alias{create_events_from_vector}
\alias{create_events_from_table}
\title{Create events from observed time series and a threshold}
\usage{
create_events_from_vector(v, threshold)

create_events_from_table(
  df,
  var,
  time_colname = "date",
  threshold_colname = NULL,
  threshold_desc = "75p_above_0.3",
  aux_vars = NULL,
  aux_funs = list(min, max, mean)
)
}
\arguments{
\item{v}{A vector (assumed to be numeric) representing some variable of
interest which has been measured at regular time intervals; Missing values
or "gaps in observed series" must be explicitly filled with NAs.}

\item{threshold}{A numeric threshold which defines the minimum value for
which an observation would be considered part of an event}

\item{df}{A data.frame of observations recorded at regular intervals (
restricted, for now, to days). A column called 'date' must exist as well as
var, specified by the user-provided argument var.}

\item{var}{(character) the variable within df for which events are defined,
e.g. "precip"}

\item{time_colname}{(character) NOT YET IMPLEMENTED the name of the column
with time information; "date" by default}

\item{threshold_colname}{(character) NOT YET IMPLEMENTED the name of the
column containing a pre-calcuated threshold; This option is provided for
users who wish to experiment with thresholds beyond what threshold_desc
allows for.}

\item{threshold_desc}{(character) NOT YET IMPLEMENTED a string of a
particular format (see )}

\item{aux_vars}{(character vector) NOT YET IMPLEMENTED other variables within
df which you want to summarize alongside events. For example, "tmin" might be
of interest during precipitation events}

\item{aux_funs}{(list of functions) NOT YET IMPLEMENTED summary functions to
to apply to aux_vars}
}
\value{
A list containing:
\itemize{
\item threshold, the vector above which the series was truncated in order to
define events
\item event_summaries_df, a data.frame() -
Each row provides a basic summary of events, including their starting
indices, durations, magnitudes, maxima, and whether or not the duration,
magnitude, and maximum are to be taken at face value}
}
\description{
Create events from observed time series and a threshold
}
\examples{
v <- c(333,587,0,201,82,225,70,62,47,101,187,NA,NA,126,222,86,NA,125)
# ^^ perhaps precipitation measured in 10ths of mm, with some missing values
#' in the mix
create_events_from_vector(v, 100)

df <- data.frame(
precip = v,
tmax = rnorm(n = 18, mean = 10, sd = 2),
date = seq.Date(from = as.Date("2025-01-01"), by = "day", length.out = 19)[-10])

create_events_from_table(df = df, var = "precip", aux_vars = "tmax")


\dontrun{
$threshold
[1] 100

$event_summaries_df
  start_index length magnitude maximum length_uncertain
1           1      2       487     487             TRUE
2           4      1       101     101            FALSE
3           6      1       125     125            FALSE
4          10      2        88      87             TRUE
5          14      2       148     122             TRUE
6          18      1        25      25             TRUE
}
}
