% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_events_from_vector.R
\name{create_events_from_vector}
\alias{create_events_from_vector}
\title{Create events from a regularly-sampled time series which is represented as a
vector}
\usage{
create_events_from_vector(v, threshold)
}
\arguments{
\item{v}{A vector (typically numeric) representing some variable of interest
which has been measured at regular time intervals.}

\item{threshold}{A numeric threshold which defines the minimum value for
which an observation would be considered part of an event}
}
\value{
unknown for now...
}
\description{
Create events from a regularly-sampled time series which is represented as a
vector
}
\examples{
v <- c(333,587,0,201,82,225) # perhaps precipitation measured in 10ths of mm
create_events_from_vector(v, 100)

\dontrun{
  start_index length magnitude maximum
1           1      2       720     487
2           4      1       101     101
3           6      1       125     125
4          10      2        88      87
}
}
