\docType{package}
\name{events}
\alias{events}
\alias{events-package}
\alias{package-events}
\title{Stores and manipulates event data}
\description{
  Stores, manipulates, scales, aggregates and creates
  directed dyadic time series from event data generated by
  KEDS, TABARI, or any other extraction tool with similarly
  structured output.
}
\details{
  Events offers simple methods for aggregating and renaming
  actors and event codes, applying event scales, and
  constructing regular time series at a choice of temporal
  scales and measurement levels.
}
\author{
  Will Lowe \email{will.lowe@uni-mannheim.de}
}

