% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtetlg.R
\name{gtetlg}
\alias{gtetlg}
\alias{rgtetlg}
\alias{pgtetlg}
\alias{pgtetlgdt}
\title{Title}
\usage{
rgtetlg(n, beta, prob_p, prob_q, rounding = FALSE)

pgtetlg(
  x_lower = (1e-10)/beta,
  x_upper = Inf,
  y_lower = (1e-10)/beta,
  y_upper = Inf,
  n_lower = 1,
  n_upper = Inf,
  beta,
  prob_p,
  prob_q,
  x_lims = NULL,
  y_lims = NULL,
  n_lims = NULL
)

pgtetlgdt(q, beta, prob_p, prob_q, lower.tail = TRUE)
}
\arguments{
\item{rounding}{}
}
\description{
Title
}
