% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtetlg.R
\name{gtetlg}
\alias{gtetlg}
\alias{rgtetlg}
\title{Title}
\usage{
rgtetlg(n, beta, prob_p, prob_q, rounding = FALSE)
}
\arguments{
\item{n}{when provided to rsmp, this is the number of observations. Otherwise
it represents the duration of an event.}

\item{beta}{numeric > 0; the rate parameter of the exponential distribution
which describes individual exceedances. It is the reciprocal of the mean.}

\item{rounding}{logical; if TRUE, generated Pareto II values are rounded to
the nearest whole number. This feature is present in case this is important
for a comparison of actual observations to some theoretical output.}
}
\description{
Title
}
