% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdiscretelomax.R
\name{hdiscretelomax}
\alias{hdiscretelomax}
\alias{dhdiscretelomax}
\title{The Hurdle Discrete}
\usage{
hgeom(x, q, p, n, prob_p, prob_q, log, log.p, lower.tail)

dhdiscretelomax(x, prob_q, dlomax_prob_p, dlomax_alpha, log = FALSE)
}
\arguments{
\item{x, q}{vector of quantiles representing the number of failures in a
sequence of (non)Bernoulli trials before success occurs.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be
the number required.}

\item{prob_q}{probability of success in the first trial. \eqn{0 \le} prob_q \eqn{\le 1}}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{prob_p}{(approximate) probability of success in each trial after the
first trial, if the first trial was not a success. \eqn{0 <} prob_p \eqn{\le 1}.}
}
\value{
hdiscretelomax is a dummy function to keep the hurdle geometric documentation
together.

dhdiscretelomax gives the density, phdiscretelomax gives the distribution
function, qhdiscretelomax, gives the quantile function, and rhdiscretelomax
generates random deviates.

Invalid dlomax_prob_p or prob_q will result in return value NaN, with a warning.
}
\description{
Density, distribution function, quantile function and random generation for
the 1-inflated discrete lomax distribution with parameters prob_q,
dlomax_alpha, and dlomax_prob_p. This is practically a hurdle geometric
distribution when dlomax_alpha is near 0. References to "bernoulli trials"
and the like here are not formally valid, but are instructive. Deviations
from the actual geometric distribution are written in parentheses.
}
