% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgsmp.R
\name{hgsmp}
\alias{hgsmp}
\alias{rhgsmp}
\title{Title}
\usage{
rhgsmp(n, alpha, beta, prob_p, prob_q, rounding = FALSE)
}
\arguments{
\item{n}{when provided to rsmp, this is the number of observations. Otherwise
it represents the duration of an event.}

\item{alpha, beta}{tail/shape parameter and scale parameters of a Multivariate
Pareto II distribution (see details for the parameterization used).}

\item{rounding}{logical; if TRUE, generated Pareto II values are rounded to
the nearest whole number. This feature is present in case this is important
for a comparison of actual observations to some theoretical output.}
}
\description{
Title
}
