% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{map_eventdata}
\alias{map_eventdata}
\title{Aggregate field values}
\usage{
map_eventdata(edo, fun, which)
}
\arguments{
\item{edo}{Event data}

\item{fun}{Function specifying the aggregation mapping, or a list}

\item{which}{Name of the field to map}
}
\value{
Event data with new fields
}
\description{
Aggregates event data fields
}
\details{
DEPRECATED This function applies a mapping/aggregration function to event data.
It is the workhorse function behind the \code{map_} functions.
You should use these in ordinary use.
<<<<<<< HEAD
=======
}
\author{
Will Lowe
>>>>>>> 04e7b17b1e0e1dc576ccecff1ee015c5e8a3d0e2
}
\seealso{
\code{\link{map_codes}}, \code{\link{map_actors}}
}
\author{
Will Lowe
}
