% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{print.eventdata}
\alias{print.eventdata}
\title{Show the first few events in an event data set}
\usage{
\method{print}{eventdata}(edo, ...)
}
\arguments{
\item{edo}{Event data}
}
\value{
Silently returns the event data set
}
\description{
Prints out the first few events of an event data set
}
\author{
Will Lowe
}
