% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_keds2}
\alias{read_keds2}
\title{Read in KEDS output files}
\usage{
read_keds2(d, one.a.day = TRUE)
}
\arguments{
\item{d}{one or more names of KEDS output}

\item{one.a.day}{whether we should apply the one-a-day event filter?}
}
\value{
an event data set
}
\description{
Read in KEDS output files
}

