\name{sources}
\alias{sources}
\title{List source actor codes}
\usage{
  sources(edo)
}
\arguments{
  \item{edo}{Event data}
}
\value{
  Array of actor codes
}
\description{
  Lists source actor codes
}
\details{
  Lists all the actor codes that appear as a source in the
  event data in alphabetical order.
}
\author{
  Will Lowe
}
\seealso{
  \code{\link{actors}}, \code{\link{targets}},
  \code{\link{codes}}
}

