% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{spotter}
\alias{spotter}
\title{Make a spotting function}
\usage{
spotter(...)
}
\arguments{
\item{...}{The actor names for which the new function should return \code{TRUE}}
}
\value{
A function
}
\description{
Hands back a function to spot the items it was given in (\code{...})
}
\details{
This is a convenience function for creates a function that returns
true for exact matches to its arguments.
}
\examples{
data("balkans.weis")
head(balkans.weis, 3)
sp <- spotter("SER", "SERMIL")
events <- filter_actors(balkans.weis, sp)
head(events, 3)
}
\author{
Will Lowe
}
