\name{summary.eventscale}
\alias{summary.eventscale}
\title{Summarise an eventscale}
\usage{
  \method{summary}{eventscale} (es)
}
\arguments{
  \item{es}{Scale}
}
\value{
  Nothing, used for side effect
}
\description{
  Summarise an eventscale
}
\details{
  Print summary statistics for an eventscale.
}
\author{
  Will Lowe
}

