% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{targets}
\alias{targets}
\title{Lists target actor codes}
\usage{
targets(edo)
}
\arguments{
\item{edo}{Event data}
}
\value{
Array of actor codes
}
\description{
Lists target actor codes
}
\details{
Lists all the actor codes that appear as a target in the event data
in alphabetical order.
}
\examples{
data(levant.cameo)
targs <- sources(levant.cameo)
head(targs)
tail(targs)
}
\seealso{
\code{\link{sources}}, \code{\link{actors}}, \code{\link{codes}}
}
\author{
Will Lowe
}
