\name{targets}
\alias{targets}
\title{Lists target actor codes}
\usage{
  targets(edo)
}
\arguments{
  \item{edo}{Event data}
}
\value{
  Array of actor codes
}
\description{
  Lists target actor codes
}
\details{
  Lists all the actor codes that appear as a target in the
  event data in alphabetical order.
}
\author{
  Will Lowe
}
\seealso{
  \code{\link{sources}}, \code{\link{actors}},
  \code{\link{codes}}
}

