\docType{data}
\name{weis.goldstein.scale}
\alias{weis.goldstein.scale}
\title{WEIS codes to Goldstein's conflict-cooperation scale}
\description{
  A mapping of WEIS event codes to [-10,10] representing a
  scale of conflict and cooperation, developed by Joshua
  Goldstein. Note: This mapping does not cover all the
  events in \code{balkans.weis}. Taken from the KEDS
  Project's documentation.
}
\author{
  KEDS Project
}
\references{
  \url{http://web.ku.edu/~keds/}
}
\keyword{data}
\keyword{datasets}

