\docType{data}
\name{weis.goldstein.scale}
\alias{weis.goldstein.scale}
\title{WEIS codes to Goldstein conflict-cooperation scale}
\description{
  A mapping of WEIS event codes to [-10,10] representing a
  scale of conflict and cooperation, developed by Joshua
  Goldstein and slightly extended for the KEDS project.
  Note: This mapping does not cover all the event codes in
  \code{balkans.weis}. Taken from the KEDS Project's
  documentation.
}
\author{
  KEDS Project
}
\references{
  \url{http://web.ku.edu/~keds/}
}
\keyword{data}
\keyword{dataset}

